% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarm_core.R
\name{icarm_calibrate}
\alias{icarm_calibrate}
\title{Probability calibration diagnostics}
\usage{
icarm_calibrate(object, data, outcome, positive = NULL, n_bins = 10L)
}
\arguments{
\item{object}{An `icarm_model` (binary only).}

\item{data}{A data frame.}

\item{outcome}{Character outcome column.}

\item{positive}{Positive class label.}

\item{n_bins}{Number of bins (default 10).}
}
\value{
An object of class `icarm_calibration`.
}
\description{
Probability calibration diagnostics
}
\examples{
m   <- icarm_fit(
  Petal.Width ~ Sepal.Length + Sepal.Width, iris,
  model = "linear")
# calibration only for binary:
data(icarm_medical)
m2  <- icarm_fit(readmitted ~ ., icarm_medical)
cal <- icarm_calibrate(m2, icarm_medical, "readmitted", "Yes")
print(cal)
}
