% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarm_core.R
\name{icarm_equity_summary}
\alias{icarm_equity_summary}
\title{Equity summary from a fairness report}
\usage{
icarm_equity_summary(fairness)
}
\arguments{
\item{fairness}{An `icarm_fairness` from [icarm_fairness()].}
}
\value{
A named list of scalar equity indicators.
}
\description{
Equity summary from a fairness report
}
