% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarm_core.R
\name{icarm_explain}
\alias{icarm_explain}
\title{Generate global model explanations}
\usage{
icarm_explain(object, data = NULL, label = NULL)
}
\arguments{
\item{object}{An `icarm_model` from [icarm_fit()].}

\item{data}{Optional data frame for DALEX explainer.}

\item{label}{Optional label for DALEX explainer.}
}
\value{
An object of class `icarm_explainer`.
}
\description{
Generate global model explanations
}
\examples{
m  <- icarm_fit(Species ~ ., iris)
ex <- icarm_explain(m)
print(ex)
print(icarm_plot_importance(ex))
}
