% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarm_core.R
\name{icarm_explain_local}
\alias{icarm_explain_local}
\title{Local explanation for individual observations}
\usage{
icarm_explain_local(explainer, newdata, n_features = 10L)
}
\arguments{
\item{explainer}{An `icarm_explainer` from [icarm_explain()].}

\item{newdata}{A data frame of observations to explain.}

\item{n_features}{Max features to show (default 10).}
}
\value{
A list of tibbles, one per row of newdata.
}
\description{
Local explanation for individual observations
}
\examples{
m  <- icarm_fit(Species ~ ., iris)
ex <- icarm_explain(m)
icarm_explain_local(ex, iris[1:2, ])
}
