% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{icarm_plot_calibration}
\alias{icarm_plot_calibration}
\title{Plot calibration curve}
\usage{
icarm_plot_calibration(calibration, title = NULL)
}
\arguments{
\item{calibration}{An `icarm_calibration` from [icarm_calibrate()].}

\item{title}{Optional title.}
}
\value{
A ggplot2 object.
}
\description{
Plot calibration curve
}
