% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{icarm_plot_confusion}
\alias{icarm_plot_confusion}
\title{Plot confusion matrix}
\usage{
icarm_plot_confusion(y_true, y_pred, title = NULL)
}
\arguments{
\item{y_true}{Factor of true outcomes.}

\item{y_pred}{Factor of predicted outcomes.}

\item{title}{Optional title.}
}
\value{
A ggplot2 object.
}
\description{
Plot confusion matrix
}
\examples{
m    <- icarm_fit(Species ~ ., iris)
yhat <- predict(m, iris)
print(icarm_plot_confusion(iris$Species, yhat))
}
