% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{icarm_plot_fairness}
\alias{icarm_plot_fairness}
\title{Plot group-level fairness metric}
\usage{
icarm_plot_fairness(fairness, metric = "acc", title = NULL, ref_line = NULL)
}
\arguments{
\item{fairness}{An `icarm_fairness` from [icarm_fairness()].}

\item{metric}{Character. Column to plot.}

\item{title}{Optional title.}

\item{ref_line}{Optional numeric reference line.}
}
\value{
A ggplot2 object.
}
\description{
Plot group-level fairness metric
}
\examples{
m    <- icarm_fit(Species ~ ., iris)
idf  <- iris
idf$size <- factor(ifelse(idf$Sepal.Length > 5.8, "large", "small"))
f <- icarm_fairness(m, idf, "Species", "size")
print(icarm_plot_fairness(f, metric = "acc"))
}
