% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarm_core.R
\name{icarm_scorecard}
\alias{icarm_scorecard}
\title{Generate a full accountability scorecard}
\usage{
icarm_scorecard(
  object,
  test_data,
  outcome,
  protected = NULL,
  positive = NULL,
  analyst = NULL,
  project = "icarm",
  path = NULL
)
}
\arguments{
\item{object}{An `icarm_model`.}

\item{test_data}{Data frame of test data.}

\item{outcome}{Character outcome column.}

\item{protected}{Optional protected attribute column.}

\item{positive}{Positive class (binary).}

\item{analyst}{Character analyst name.}

\item{project}{Character project name.}

\item{path}{Optional JSON output path.}
}
\value{
Invisibly, the scorecard list.
}
\description{
Generate a full accountability scorecard
}
\examples{
sp <- icarm_split(iris, stratify = "Species")
m  <- icarm_fit(Species ~ ., sp$train)
iris_test <- sp$test
iris_test$size <- factor(
  ifelse(iris_test$Sepal.Length > 5.8, "large","small"))
icarm_scorecard(m, iris_test, outcome="Species",
                protected="size", project="Iris Demo")
}
