% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarm_fit.R
\name{predict.icarm_model}
\alias{predict.icarm_model}
\title{Predict from an icarm_model}
\usage{
\method{predict}{icarm_model}(object, newdata, type = c("class", "prob"), threshold = 0.5, ...)
}
\arguments{
\item{object}{An `icarm_model`.}

\item{newdata}{A data frame for prediction.}

\item{type}{For classification: `"class"` or `"prob"`.
For regression: ignored.}

\item{threshold}{Decision threshold for binary (default 0.5).}

\item{...}{Ignored.}
}
\value{
Factor vector, probability matrix, or numeric vector.
}
\description{
Predict from an icarm_model
}
\examples{
m <- icarm_fit(Species ~ ., iris)
predict(m, iris[1:5, ], type = "class")
}
