% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarm_fit.R
\name{print.icarm_model}
\alias{print.icarm_model}
\title{Print an icarm_model}
\usage{
\method{print}{icarm_model}(x, ...)
}
\arguments{
\item{x}{An icarm_model object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the icarm_model object x. Called for
  its side effect of printing a formatted summary to the console.
}
\description{
Print an icarm_model
}
