% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarm_fit.R
\name{summary.icarm_model}
\alias{summary.icarm_model}
\title{Summary of an icarm_model}
\usage{
\method{summary}{icarm_model}(object, ...)
}
\arguments{
\item{object}{An icarm_model object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the summary of the underlying fitted
  model object. Called for its side effect of printing a detailed
  model summary to the console.
}
\description{
Summary of an icarm_model
}
